pyproject.toml
[project]
name = "pick-twitter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.13.0",
    "datasets>=4.8.4",
    "huggingface-hub[cli]>=1.9.0",
    "openai>=2.30.0",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "torch>=2.11.0",
    "transformers>=5.5.0",
]
